feat: отправить в третью итерацию ревью, первую итерацию автотестов
